TestsPrinter: use parametrizedArrayParameters in printFunctionParameters
Use TestsPrinter::parametrizedArrayParameters method in
TestsPrinter::printFunctionParameters and call it for every parameter.
I decided to do it because TestsPrinter::parametrizedArrayParameters
also outputs function parameters.

This change fixes the appearance of two structs in generated tests.
The point of thing is that if we pass to a function some pointer to
a struct:
	bool OKrow(int x, const sudoku *a, int y, int v);
where:
	struct sudoku
	{
    	    unsigned int *a;
    	    unsigned int N;
    	    unsigned int N2;
	};

We have the pointer to the struct and this struct contains a lazy value.

Refs: #475