TestsPrinters.cpp: add FAIL() assertion to error tests.
Add FAIL() assertion to generated tests in no verbose mode.
The line with FAIL() assertion is unreachable in some tests but it gives
additional information to the user to make the tests easier to read.

I add a unit test to check that the code of generated tests (error suit) has
the line with FAIL() assertion. I decided not to use statuses from the tests
execution because the line with FAIL() assertion can be unreachable.

Refs: #472